# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Movie Library API",
    "version": "1.0.0",
    "description": "CRUD API for managing movies and reviews with OAuth authentication"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Local Server"
    },
    {
      "url": "https://movies-library-api-1.onrender.com",
      "description": "Production Server"
    }
  ],
  "paths": {
    "/movies": {
      "get": {
        "summary": "Get all movies",
        "responses": {
          "200": {
            "description": "Movies retrieved successfully"
          }
        }
      },
      "post": {
        "summary": "Create a movie (Protected)",
        "security": [
          {
            "githubAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "title": {
                    "type": "string"
                  },
                  "director": {
                    "type": "string"
                  },
                  "genre": {
                    "type": "string"
                  },
                  "releaseYear": {
                    "type": "number"
                  },
                  "rating": {
                    "type": "number"
                  },
                  "duration": {
                    "type": "number"
                  },
                  "language": {
                    "type": "string"
                  },
                  "streamingPlatform": {
                    "type": "string"
                  }
                }
              },
              "example": {
                "title": "Interstellar",
                "director": "Christopher Nolan",
                "genre": "Sci-Fi",
                "releaseYear": 2014,
                "rating": 8.6,
                "duration": 169,
                "language": "English",
                "streamingPlatform": "Netflix"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Movie created"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },

    "/movies/{id}": {
      "get": {
        "summary": "Get movie by ID",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Movie found"
          }
        }
      },

      "put": {
        "summary": "Update movie (Protected)",
        "security": [
          {
            "githubAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "example": {
                "title": "Updated Movie"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Movie updated"
          }
        }
      },

      "delete": {
        "summary": "Delete movie (Protected)",
        "security": [
          {
            "githubAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Movie deleted"
          }
        }
      }
    }
  },

  "components": {
    "securitySchemes": {
      "githubAuth": {
        "type": "oauth2",
        "flows": {
          "authorizationCode": {
            "authorizationUrl": "https://github.com/login/oauth/authorize",
            "tokenUrl": "https://github.com/login/oauth/access_token",
            "scopes": {}
          }
        }
      }
    }
  }
}